Moved .NET standard back office identity code into infrastructure (lowest common dependency)
