+++ src/Umbraco.Infrastructure/Umbraco.Infrastructure.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>netstandard2.0</TargetFramework>
@@ -16,6 +16,7 @@
       <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.2" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.2" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.2" />
+      <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="3.1.2" />
       <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" />
       <PackageReference Include="MiniProfiler.Shared" Version="4.1.0" />
       <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
